db/conexao.py
import pyodbc
import os
from dotenv import load_dotenv

load_dotenv()

def conectar():
    conn = pyodbc.connect(
        f"DRIVER={{{os.getenv('DB_DRIVER')}}};"
        f"SERVER={os.getenv('DB_SERVER')};"
        f"DATABASE={os.getenv('DB_NAME')};"
        "Trusted_Connection=yes;"
    )
    return conn
