requirements.txt
requests
beautifulsoup4
pandas
cloudscraper
python-dotenv